+++ test/test_algo_ag.py
@@ -8,6 +8,46 @@ import pytest
 from wordseg.algos import ag
 
 
+# # this test is not stable enough, so it is commented out
+# def test_replicate_cdswordseg(datadir, tmpdir):
+#     sep = Separator()
+
+#     _tags = [utt for utt in codecs.open(
+#         os.path.join(datadir, 'tagged.txt'), 'r', encoding='utf8')
+#             if utt][:200]
+#     _prepared = list(prepare(_tags, separator=sep))
+#     _gold = gold(_tags, separator=sep)
+
+#     # build a colloc0 grammar from the list of phones
+#     phones = {p for utt in _prepared for p in utt.split(' ')}
+#     grammar = ag.build_colloc0_grammar(phones)
+#     grammar_file = tmpdir.join('grammar.lt')
+#     grammar_file.write_text(grammar, encoding='utf8')
+
+#     segmented = ag.segment(_prepared, str(grammar_file), 'Colloc0',
+#                            # log=utils.get_logger(level=logging.DEBUG),
+#                            args=test_arguments.replace('-n 10', '-n 2'),
+#                            nruns=1)
+#     score = evaluate(segmented, _gold)
+
+#     # we obtained those results from the AG version in CDSwordseg
+#     # (preparing with wordseg-prep and evaluating with wordseg-eval)
+#     expected = {
+#         'type_fscore': 0.2385,
+#         'type_precision': 0.2038,
+#         'type_recall': 0.2875,
+#         'token_fscore': 0.2642,
+#         'token_precision': 0.238,
+#         'token_recall': 0.2969,
+#         'boundary_fscore': 0.5361,
+#         'boundary_precision': 0.4754,
+#         'boundary_recall': 0.6145}
+
+#     # Because we use so few data and iterations,  the results are
+#     # not stable, we limit the precision to 0.3
+#     for k, v in score.items():
+#         assert expected[k] == pytest.approx(v, abs=0.3), k
+
 TEST_ARGUMENTS = (
     '-E -P -R -1 -n 10 -a 0.0001 -b 10000.0 '
     '-e 1.0 -f 1.0 -g 100.0 -h 0.01 -x 2')
@@ -115,7 +155,9 @@ def test_parse_counter(njobs):
 @pytest.mark.parametrize('grammar, level', GRAMMARS)
 def test_grammars(prep, grammar, level):
     grammar = os.path.join(GRAMMAR_DIR, grammar)
-    segmented = ag.segment(prep, grammar, level, TEST_ARGUMENTS, nruns=1)
+    segmented = ag.segment(
+        prep, grammar_file=grammar, category=level,
+        args=TEST_ARGUMENTS, nruns=1)
     assert len(segmented) == len(prep)
 
     segmented = ''.join(utt.replace(' ', '').strip() for utt in segmented)
@@ -147,57 +189,63 @@ def test_mark_jonhson(tmpdir, datadir):
         # -U "cat > {prs1}" -v {testeng2} -V "cat > {prs2}"'
         '-u {testeng1} '.format(
             testeng1=os.path.join(datadir, 'ag_testeng1.yld'),
-            testeng2=os.path.join(datadir, 'ag_testeng2.yld'),
+            # testeng2=os.path.join(datadir, 'ag_testeng2.yld'),
             trace=tmpdir.join('trace'), wlt=tmpdir.join('wlt'),
             prs=tmpdir.join('prs')
             # X1=tmpdir.join('X1'), X2=tmpdir.join('X2'),
             # prs1=tmpdir.join('prs1'), prs2=tmpdir.join('prs2')))
         ))
-    pc = ag.ParseCounter(len(text))
-    output = ag.segment(text, grammar_file, category='VP', args=arguments,
-                        ignore_first_parses=0, nruns=1)
+    # pc = ag.ParseCounter(len(text))
+    output = ag.segment(text, grammar_file=grammar_file, category='VP',
+                        args=arguments, ignore_first_parses=0, nruns=1)
     assert len(text) == len(output)
     for i in range(len(text)):
         assert text[i].strip().replace(' ', '') == output[i].replace(' ', '')
 
 
+# AG: <train_text is None>
+@pytest.mark.parametrize('grammar, level', GRAMMARS)
+def test_traintext_None(prep, grammar, level):
+    grammar = os.path.join(GRAMMAR_DIR, grammar)
+    segmented = ag.segment(
+        prep, train_text=None, grammar_file=grammar, category=level,
+        args=TEST_ARGUMENTS, nruns=1)
+    assert len(segmented) == len(prep)
 
-# # this test is not stable enough, so it is commented out
-# def test_replicate_cdswordseg(datadir, tmpdir):
-#     sep = Separator()
+    segmented = ''.join(utt.replace(' ', '').strip() for utt in segmented)
+    prep = ''.join(utt.replace(' ', '').strip() for utt in prep)
+    assert segmented == prep
 
-#     _tags = [utt for utt in codecs.open(
-#         os.path.join(datadir, 'tagged.txt'), 'r', encoding='utf8')
-#             if utt][:200]
-#     _prepared = list(prepare(_tags, separator=sep))
-#     _gold = gold(_tags, separator=sep)
 
-#     # build a colloc0 grammar from the list of phones
-#     phones = {p for utt in _prepared for p in utt.split(' ')}
-#     grammar = ag.build_colloc0_grammar(phones)
-#     grammar_file = tmpdir.join('grammar.lt')
-#     grammar_file.write_text(grammar, encoding='utf8')
+# AG: <train_text=test_text>
+@pytest.mark.parametrize('grammar, level', GRAMMARS)
+def test_traintext_equal_testtext(prep, grammar, level):
+    grammar = os.path.join(GRAMMAR_DIR, grammar)
+    segmented = ag.segment(
+        prep, train_text=prep, grammar_file=grammar, category=level,
+        args=TEST_ARGUMENTS, nruns=1)
+    assert len(segmented) == len(prep)
 
-#     segmented = ag.segment(_prepared, str(grammar_file), 'Colloc0',
-#                            # log=utils.get_logger(level=logging.DEBUG),
-#                            args=test_arguments.replace('-n 10', '-n 2'),
-#                            nruns=1)
-#     score = evaluate(segmented, _gold)
+    segmented = ''.join(utt.replace(' ', '').strip() for utt in segmented)
+    prep = ''.join(utt.replace(' ', '').strip() for utt in prep)
+    assert segmented == prep
 
-#     # we obtained those results from the AG version in CDSwordseg
-#     # (preparing with wordseg-prep and evaluating with wordseg-eval)
-#     expected = {
-#         'type_fscore': 0.2385,
-#         'type_precision': 0.2038,
-#         'type_recall': 0.2875,
-#         'token_fscore': 0.2642,
-#         'token_precision': 0.238,
-#         'token_recall': 0.2969,
-#         'boundary_fscore': 0.5361,
-#         'boundary_precision': 0.4754,
-#         'boundary_recall': 0.6145}
 
-#     # Because we use so few data and iterations,  the results are
-#     # not stable, we limit the precision to 0.3
-#     for k, v in score.items():
-#         assert expected[k] == pytest.approx(v, abs=0.3), k
+# AG: <train_text!=test_text>
+@pytest.mark.parametrize('grammar, level', GRAMMARS)
+def test_traintext_notequal_testtext(grammar, level):
+    # hello world
+    train_text = ['hh ax l ow w er l d'] * 10
+
+    # good morn, that dog is big (no phones shared with train)
+    test_text = ['g uh d m ao r n', 'dh ae t d ao g ih z b ih g']
+
+    grammar = os.path.join(GRAMMAR_DIR, grammar)
+    segmented = ag.segment(
+        test_text, train_text=train_text, grammar_file=grammar, category=level,
+        args=TEST_ARGUMENTS, nruns=1)
+    assert len(segmented) == len(test_text)
+
+    segmented = ''.join(utt.replace(' ', '').strip() for utt in segmented)
+    prep = ''.join(utt.replace(' ', '').strip() for utt in test_text)
+    assert segmented == prep