+++ test/test_algo_dibs.py
@@ -32,7 +32,7 @@ def test_basic(prep, tags, type, threshold, pwb):
 
 @pytest.mark.parametrize('level', ['phone', 'syllable'])
 def test_phone_sep(level):
-    text = ['hh_ih_r_;eword ',
+    test_text = ['hh_ih_r_;eword ',
             'dh_eh_r_;eword w_iy_;eword g_ow_;eword ']
 
     sep = Separator(
@@ -40,7 +40,7 @@ def test_phone_sep(level):
         syllable='_' if level == 'syllable' else None,
         word=';eword ')
 
-    model = dibs.CorpusSummary(text, separator=sep, level=level)
+    model = dibs.CorpusSummary(test_text, separator=sep, level=level)
     assert model.summary == {'nlines': 2, 'nwords': 4, 'nphones': 10}
 
 
@@ -49,12 +49,12 @@ def test_bad_train(prep):
     with pytest.raises(ValueError):
         dibs.CorpusSummary(prep)
 
-
+#Dibs: <train_text is None>
 def test_emptyline_in_train(tags):
     # empty lines are ignored in train file
     dibs.CorpusSummary([''] + tags)
 
-
+#Dibs: <train_text is not None>
 def test_replicate_cdswordseg(datadir):
     sep = Separator()
 
@@ -88,3 +88,4 @@ def test_replicate_cdswordseg(datadir):
         'boundary_noedge_recall': 0.3614}
 
     assert score == pytest.approx(expected, rel=1e-3)
+