+++ wordseg/algos/tp.py
@@ -1,9 +1,11 @@
 """Transitional Probabilities word segmentation"""
 
-# Author: Amanda Saksida, Mathieu Bernard
+# Author: Amanda Saksida, Mathieu Bernard, Manel Khentout
 
+import codecs
 import collections
 import math
+import os
 import re
 
 from wordseg import utils
@@ -54,16 +56,66 @@ def _threshold_absolute(units, tps):
     return cwords
 
 
-def segment(text, threshold='relative', dependency='ftp',
+# -----------------------------------------------------------------------------
+#  Training
+# -----------------------------------------------------------------------------
+
+def _train(train_units, dependency):
+    # compute and count all the unigrams and bigrams (two successive units)
+    unigrams = collections.Counter(train_units)
+    bigrams = collections.Counter(zip(train_units[0:-1], train_units[1:]))
+
+    # compute the transitional probabilities accordoing to the given
+    # dependency measure
+
+    if dependency == 'ftp':
+        tps = collections.defaultdict(lambda: 0, {
+            bigram: float(freq) / unigrams[bigram[0]]
+            for bigram, freq in bigrams.items()})
+    elif dependency == 'btp':
+        tps = collections.defaultdict(lambda: 0, {
+            bigram: float(freq) / unigrams[bigram[1]]
+            for bigram, freq in bigrams.items()})
+    else:  # dependency == 'mi'
+        tps = collections.defaultdict(lambda: 0, {
+            bigram: math.log(float(freq) / (
+                unigrams[bigram[0]] * unigrams[bigram[1]]), 2)
+            for bigram, freq in bigrams.items()})
+    return tps
+
+
+# -----------------------------------------------------------------------------
+#  Segmentation
+# -----------------------------------------------------------------------------
+
+def _segment(units, tps, threshold):
+    # segment the input given the transition probalities
+    cwords = (_threshold_relative(units, tps) if threshold == 'relative'
+              else _threshold_absolute(units, tps))
+    segtext = ' '.join(''.join(c) for c in cwords)
+    return [utt.strip() for utt in re.sub(' +', ' ', segtext).split('UB')]
+
+
+# -----------------------------------------------------------------------------
+#  Segment function
+# -----------------------------------------------------------------------------
+
+def segment(text, train_text=None, threshold='relative', dependency='ftp',
             log=utils.null_logger()):
     """Returns a word segmented version of `text` using the TP algorithm
+
+    The parameters `text` and `train_text` must be formatted as follows: A
+        sequence of lines with syllable (or phoneme) boundaries marked by
+        spaces and no word boundaries. Each line in the sequence corresponds to
+        a single and complete utterance
 
     Parameters
     ----------
     text : sequence
-        A sequence of lines with syllable (or phoneme) boundaries
-        marked by spaces and no word boundaries. Each line in the
-        sequence corresponds to a single and complete utterance.
+        The text to segment into words
+    train_text : sequence, optional
+        The text used to train model on (estimation of transition
+        probabilities). If not specified use the `text`.
     threshold : str, optional
         Type of threshold to use, must be 'relative' or 'absolute'.
     dependency : str, optional
@@ -86,7 +138,7 @@ def segment(text, threshold='relative', dependency='ftp',
 
     """
     # raise on invalid threshold type
-    if threshold != 'relative' and threshold != 'absolute':
+    if threshold not in ('relative', 'absolute'):
         raise ValueError(
             "invalid threshold, must be 'relative' or 'absolute', it is '{}'"
             .format(threshold))
@@ -100,36 +152,24 @@ def segment(text, threshold='relative', dependency='ftp',
     log.info('running TP with %s threshold and %s dependency measure',
              threshold, dependency)
 
-    # join all the utterances together, seperated by ' UB '
-    units = [unit for unit in ' UB '.join(
-        line.strip() for line in text).split()]
+    # calculate test_unit and train_unit
+    test_units = ' UB '.join(line.strip() for line in text).split()
 
-    # compute and count all the unigrams and bigrams (two successive units)
-    unigrams = collections.Counter(units)
-    bigrams = collections.Counter(zip(units[0:-1], units[1:]))
+    if train_text is None:
+        train_units = test_units
+    else:
+        train_units = ' UB '.join(line.strip() for line in train_text).split()
 
-    # compute the transitional probabilities accordoing to the given
-    # dependency measure
-    if dependency == 'ftp':
-        tps = {bigram: float(freq) / unigrams[bigram[0]]
-               for bigram, freq in bigrams.items()}
-    elif dependency == 'btp':
-        tps = {bigram: float(freq) / unigrams[bigram[1]]
-               for bigram, freq in bigrams.items()}
-    else:  # dependency == 'mi'
-        tps = {bigram: math.log(float(freq) / (
-            unigrams[bigram[0]] * unigrams[bigram[1]]), 2)
-               for bigram, freq in bigrams.items()}
+    # estimate the transition probabilities
+    tps = _train(train_units, dependency)
 
-    # segment the input given the transition probalities
-    cwords = (_threshold_relative(units, tps) if threshold == 'relative'
-              else _threshold_absolute(units, tps))
+    # segment the text using those TPs
+    return _segment(test_units, tps, threshold)
 
-    # format the segment text for output (' UB ' -> '\n', remove
-    # multiple spaces)
-    segtext = ' '.join(''.join(c) for c in cwords)
-    return [utt.strip() for utt in re.sub(' +', ' ', segtext).split('UB')]
 
+# -----------------------------------------------------------------------------
+#  Command line arguments
+# -----------------------------------------------------------------------------
 
 def _add_arguments(parser):
     """Add algorithm specific options to the parser"""
@@ -163,10 +203,12 @@ def _add_arguments(parser):
 def main():
     """Entry point of the 'wordseg-tp' command"""
     # command initialization
-    streamin, streamout, separator, log, args = utils.prepare_main(
+
+    streamin, streamout, _, log, args = utils.prepare_main(
         name='wordseg-tp',
         description=__doc__,
-        add_arguments=_add_arguments)
+        add_arguments=_add_arguments,
+        train_file=True)
 
     # if the deprecated --probability option is used, raise a warning
     # and convert it to the new --dependency option.
@@ -180,9 +222,23 @@ def main():
         else:  # 'backward'
             args.dependency = 'btp'
 
-    # segment the input text
+    # load the train text if any
+    train_text = None
+    if args.train_file:
+        if not os.path.isfile(args.train_file):
+            raise RuntimeError(
+                'test file not found: {}'.format(args.train_file))
+        train_text = codecs.open(args.train_file, 'r', encoding='utf8')
+
+    # load train and test texts, ignore empty lines
+    test_text = (line for line in streamin if line)
+    if train_text:
+        train_text = (line for line in train_text if line)
+
+    # segment the input text with the train text
     text = segment(
-        streamin,
+        test_text,
+        train_text=train_text,
         threshold=args.threshold,
         dependency=args.dependency,
         log=log)